+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/RandomWrapperRandomTest.java
@@ -15,12 +15,8 @@
 // ============================================================================
 package io.github.pr0methean.betterrandom.prng;
 
-import com.google.common.collect.ImmutableMap;
-import io.github.pr0methean.betterrandom.TestUtils;
-import io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator;
 import io.github.pr0methean.betterrandom.seed.SeedException;
-import io.github.pr0methean.betterrandom.seed.SeedGenerator;
-import java.lang.reflect.InvocationTargetException;
+import java.util.Map;
 import java.util.Random;
 import org.testng.annotations.Test;
 
@@ -34,16 +30,10 @@ public class RandomWrapperRandomTest extends BaseRandomTest {
     return RandomWrapper.class;
   }
 
-  @Override public void testAllPublicConstructors()
-      throws SeedException, IllegalAccessException, InstantiationException,
-      InvocationTargetException {
-    final BaseRandom basePrng = createRng();
-    final int seedLength = getNewSeedLength(basePrng);
-    TestUtils.testAllPublicConstructors(getClassUnderTest(),
-        ImmutableMap.<Class<?>, Object>of(int.class, seedLength, long.class, TEST_SEED,
-            byte[].class, DefaultSeedGenerator.DEFAULT_SEED_GENERATOR.generateSeed(seedLength),
-            SeedGenerator.class, DefaultSeedGenerator.DEFAULT_SEED_GENERATOR, Random.class,
-            new Random()), VERIFY_NEXT_INT_NO_CRASH);
+  @Override public Map<Class<?>, Object> constructorParams() {
+    Map<Class<?>, Object> params = super.constructorParams();
+    params.put(Random.class, new Random());
+    return params;
   }
 
   /**