+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/prng/ReseedingThreadLocalRandomWrapperTest.java
@@ -1,0 +1,27 @@
+package io.github.pr0methean.betterrandom.prng;
+
+import io.github.pr0methean.betterrandom.seed.DefaultSeedGenerator;
+import io.github.pr0methean.betterrandom.seed.FailingSeedGenerator;
+import io.github.pr0methean.betterrandom.seed.SeedException;
+import io.github.pr0methean.betterrandom.seed.SeedGenerator;
+import java.io.Serializable;
+import java.util.Map;
+import java.util.Random;
+import java.util.function.Supplier;
+
+public class ReseedingThreadLocalRandomWrapperTest extends ThreadLocalRandomWrapperTest {
+
+  @Override public void testWrapLegacy() throws SeedException {
+    ReseedingThreadLocalRandomWrapper
+        .wrapLegacy(Random::new, DefaultSeedGenerator.DEFAULT_SEED_GENERATOR).nextInt();
+  }
+
+  @Override protected Class<? extends BaseRandom> getClassUnderTest() {
+    return super.getClassUnderTest();
+  }
+
+  @Override protected BaseRandom createRng() throws SeedException {
+    return new ReseedingThreadLocalRandomWrapper(FailingSeedGenerator.FAILING_SEED_GENERATOR,
+        (Serializable & Supplier<BaseRandom>) MersenneTwisterRandom::new);
+  }
+}